use files() api in is_resource()
The files() api makes the identification of resources blury. Here we
re-implement is_resource() by mirroring the previous logic to the best
extent we can. This is not fully correct, as the files() api is not
fully compatible with the legacy api, but it is pretty close and as
correct as we can get. The semantics for what constitutes resources have
always been blurry in the first place.

Signed-off-by: Filipe Laíns <lains@riseup.net>