Removed reliance on on-disk fixtures.
Instead, generate on-disk and zip file fixtures on demand in temporary directories.

This approach makes it easier to compose different scenarios and mix-in behavior for zip files. It also makes the disk and zip scenarios more symmetric.

Re-wrote `test_reader` to utilize these new fixtures instead of assuming the files were present on disk in the test suite.